Separate cobbler IP from compass IP in cobbler snippets
Compass IP now is isolated so that during PXE process,
a pxe-server IP is available and DHCP will have its own
subnet and cobbler server owns the vip for that subnet.
This only partially closes bug #1393957, further change
will be maide in compass-core. So no "Closes-Bug" will
be added in the message.
Partial-bug: #1393957
Change-Id: I0d693f2958a33b5a093728dccf063e2114c7406b